# src/app/swagger.yaml
openapi: 3.0.2
info:
  title: Store Management API
  version: 1.2
  description: API for managing products, clients, images, and suppliers in the store.
servers:
  - url: /api/v1
paths:
...
  /clients/:
    get:
      summary: Get all clients with optional pagination
      tags:
        - Clients
      parameters:
        - name: limit
          in: query
          description: The number of clients to return
          required: false
          schema:
            type: integer
            example: 10
        - name: offset
          in: query
          description: The number of clients to skip before starting to return results
          required: false
          schema:
            type: integer
            example: 0
      responses:
        200:
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        500:
          description: Internal server error
    post:
      summary: Create a new client
      tags:
        - Clients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_name:
                  type: string
                client_surname:
                  type: string
                birthday:
                  type: string
                gender:
                  type: string
                address_id:
                  type: string
              required:
                - client_name
                - client_surname
                - birthday
                - gender
                - address_id
      responses:
        201:
          description: Client created
...
  /clients/{client_id}:
    patch:
      summary: Update client address
      tags:
        - Clients
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address_id:
                  type: string
              required:
                - address_id
      responses:
        200:
          description: Client updated
        404:
          description: Client not found
    delete:
      summary: Delete a client
      tags:
        - Clients
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Client deleted
        404:
          description: Client not found
...